Change API: Remove user ID from notes path

+++ src/main/java/link/timon/tutorial/securerest/notes/service/NoteService.java
@@ -3,6 +3,7 @@ package link.timon.tutorial.securerest.notes.service;
 import java.util.Collection;
 import java.util.Optional;
 import link.timon.tutorial.securerest.notes.common.EntityNotFoundException;
+import link.timon.tutorial.securerest.notes.common.UnauthorizedException;
 import link.timon.tutorial.securerest.notes.domain.Note;
 import link.timon.tutorial.securerest.notes.domain.User;
 import link.timon.tutorial.securerest.notes.domain.dto.NoteView;
@@ -24,20 +25,25 @@ public class NoteService {
     private final UserService userService;
 
     /**
-     * Creates a new Note for given user.
+     * Creates a new Note for authenticated user.
      *
-     * @param userId The Id off the user who created the note.
      * @param noteView The NoteView to create.
      *
      * @return The created NoteView.
      */
-    public Optional<NoteView> create(String userId, NoteView noteView) {
-        User user = userService.getCurrentUserAuthorized(userId);
+    public Optional<NoteView> create(NoteView noteView) {
         Note note = ViewMapper.INSTANCE.noteViewToModel(noteView);
 
-        note.setAuthor(user);
+        Optional<User> authenticatedUser = userService.getAuthenticatedUser();
 
-        updateUserNotes(user, note);
+        if (authenticatedUser.isEmpty()) {
+            throw new UnauthorizedException("You must be logged in to create a new note.");
+        }
+
+
+        note.setAuthor(authenticatedUser.get());
+
+        updateUserNotes(authenticatedUser.get(), note);
 
         return Optional.ofNullable(ViewMapper.INSTANCE.noteToView(noteRepository.save(note)));
     }
@@ -45,14 +51,11 @@ public class NoteService {
     /**
      * Updates the Note of the given user.
      *
-     * @param userId The user who owns the Note.
      * @param noteView The NoteView to update.
      *
      * @return The updated NoteView
      */
-    public Optional<NoteView> update(String userId, NoteView noteView) {
-        userService.checkCurrentuserAuthorized(userId);
-
+    public Optional<NoteView> update(NoteView noteView) {
         Note note = ViewMapper.INSTANCE.noteViewToModel(noteView);
         Note saved = noteRepository.save(note);
 
@@ -62,45 +65,67 @@ public class NoteService {
     /**
      * Finds all NoteViews of a given user.
      *
-     * @param userId The user to search for.
-     *
      * @return All found NoteViews, can be empty.
      */
-    public Collection<NoteView> findAllForUser(String userId) {
-        User user = userService.getCurrentUserAuthorized(userId);
+    public Collection<NoteView> findAllForAuthorizedUser() {
+        Optional<User> authenticatedUser = userService.getAuthenticatedUser();
 
-        return ViewMapper.INSTANCE.notesToViews(noteRepository.findByAuthor(user));
+        if (authenticatedUser.isEmpty()) {
+            throw new UnauthorizedException("Please login to view your notes.");
+        }
+
+        return ViewMapper.INSTANCE.notesToViews(noteRepository.findByAuthor(authenticatedUser.get()));
     }
 
     /**
-     * Searches for a specific note of given user.
+     * Searches for a specific note of authenticated user.
      *
-     * @param userId The Id of the User to search for.
      * @param noteId The Id of the Note to search.
      *
      * @return The found NoteView.
      */
-    public Optional<NoteView> findById(String userId, String noteId) {
+    public Optional<NoteView> findById(String noteId) {
         Optional<Note> note = noteRepository.findById(noteId);
+        Optional<User> authenticatedUser = userService.getAuthenticatedUser();
 
         if (note.isEmpty()) {
             throw new EntityNotFoundException(String.format("Could not find note with id=%s", noteId));
+        }
 
+        if (authenticatedUser.isEmpty() || !authenticatedUser.get().getId().equals(note.get().getAuthor().getId())) {
+            throw new UnauthorizedException("You are not allowed to view this note!");
         }
 
         return Optional.ofNullable(ViewMapper.INSTANCE.noteToView(note.get()));
     }
 
     /**
-     * Deletes a Note of given User.
+     * Deletes a Note of authenticated User.
      *
-     * @param userId The Id of the User to delete the note from.
      * @param noteId The Id of the Note to delete.
      */
-    public void delete(String userId, String noteId) {
-        User user = userService.getCurrentUserAuthorized(userId);
-        user.getNotes().removeIf((n) -> n.getId().equalsIgnoreCase(noteId));
-        userService.save(user);
+    public void delete(String noteId) {
+        Optional<User> authenticatedUser = userService.getAuthenticatedUser();
+
+        if (authenticatedUser.isEmpty()) {
+            throw new UnauthorizedException("You must be authenticated to delete a note.");
+        }
+
+        Optional<Note> noteOptional = authenticatedUser.get().getNotes()
+                .stream()
+                .filter(note -> note.getId().equals(noteId))
+                .findFirst();
+
+        if (noteOptional.isEmpty()) {
+            throw new EntityNotFoundException(String.format("Note with id=%s not found", noteId));
+        }
+
+        if (!noteOptional.get().getAuthor().getId().equals(authenticatedUser.get().getId())) {
+            throw new UnauthorizedException("You are not allowed to remove this note!");
+        }
+
+        authenticatedUser.get().getNotes().removeIf((n) -> n.getId().equalsIgnoreCase(noteId));
+        userService.save(authenticatedUser.get());
 
         noteRepository.deleteById(noteId);
     }
